use a parametrized fixture instead of an ABC
